+++ packages/plugin-divider/src/toolbar/insert-buttons.jsx
@@ -1,4 +1,4 @@
-import { InsertPluginIcon } from '../../statics/icons';
+import { InsertPluginIcon } from '../icons';
 import { DEFAULTS } from '../constants';
 
 export default ({ helpers, t }) => {
@@ -7,7 +7,7 @@ export default ({ helpers, t }) => {
       name: 'Divider',
       tooltipText: t('DividerPlugin_InsertButton_Tooltip'),
       Icon: InsertPluginIcon,
-      data: DEFAULTS,
+      componentData: DEFAULTS,
       helpers,
       t
     }