+++ packages/plugin-divider/statics/icons/index.js
@@ -1,4 +1,0 @@
-// @create-index
-
-export { default as InsertPluginIcon } from './InsertPluginIcon.js';
-