+++ packages/plugin-gallery/src/components/gallery-controls/gallery-image-settings.js
@@ -15,7 +15,7 @@ import {
   SettingsPanelFooter,
   FocusManager
 } from 'wix-rich-content-common';
-import { BackIcon } from '../../../statics/icons';
+import { BackIcon } from '../../icons';
 import styles from '../../../statics/styles/gallery-image-settings.scss';
 import GallerySettingsMobileHeader from './gallery-settings-mobile-header';
 import { isUndefined } from 'util';