+++ packages/plugin-gallery/src/gallery-component.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import isEqual from 'lodash/isEqual';
 import { GalleryViewer, getDefault } from './gallery-viewer';
 
 //eslint-disable-next-line no-unused-vars
@@ -17,7 +18,10 @@ class GalleryComponent extends React.PureComponent {
   }
 
   componentWillReceiveProps(nextProps) {
-    this.setState(this.stateFromProps(nextProps));
+    const { componentData, componentState } = this.props;
+    if (!isEqual(componentData, nextProps.componentData) || !isEqual(componentState, nextProps.componentState)) {
+      this.setState(this.stateFromProps(nextProps));
+    }
   }
 
   stateFromProps = props => {
@@ -63,6 +67,15 @@ class GalleryComponent extends React.PureComponent {
       items = [...items];
       items[itemPos] = item;
     }
+
+    //when updating componentData on an async method like this one,
+    // we need to use a sync method to change the EditorState.
+    // The broadcast is good if the toolbar is displaying some status or images
+    this.props.componentData.items = items;
+    this.props.componentData.styles = styles;
+    const { setData } = this.props.blockProps;
+    setData(this.props.componentData);
+
     this.setState({ items });
     if (this.props.store) {
       this.props.store.update('componentData', { items, styles, config: {} });
@@ -122,7 +135,7 @@ class GalleryComponent extends React.PureComponent {
     } else {
       handleFileAdded(data, itemIdx);
     }
-  }
+  };
 
   fileLoaded = (event, file, itemPos) => {