+++ packages/plugin-gallery/src/toolbar/insert-buttons.jsx
@@ -1,5 +1,5 @@
 import { getDefault } from './../gallery-component';
-import { InsertPluginIcon } from '../../statics/icons';
+import { InsertPluginIcon } from '../icons';
 
 export default ({ helpers, t }) => {
   return [
@@ -9,7 +9,7 @@ export default ({ helpers, t }) => {
       name: 'Gallery',
       tooltipText: t('GalleryPlugin_InsertButton_Tooltip'),
       Icon: InsertPluginIcon,
-      data: getDefault(),
+      componentData: getDefault(),
       helpers,
     },
   ];