+++ packages/plugin-html/src/toolbar/insertButtons.js
@@ -1,5 +1,5 @@
 import { DEFAULTS } from '../HtmlComponent';
-import { InsertPluginIcon } from '../../statics/icons';
+import { InsertPluginIcon } from '../icons';
 
 export default ({ helpers, t }) => {
   return [
@@ -7,7 +7,7 @@ export default ({ helpers, t }) => {
       name: 'HTML',
       tooltipText: t('HtmlPlugin_InsertButton_Tooltip'),
       Icon: InsertPluginIcon,
-      data: DEFAULTS,
+      componentData: DEFAULTS,
       helpers,
     },
   ];