+++ packages/plugin-video/src/toolbar/insert-buttons.jsx
@@ -1,7 +1,7 @@
 import { DEFAULTS } from '../video-component';
 import { getModalStyles } from 'wix-rich-content-common';
 import VideoURLInputModal from './videoURLInputModal';
-import { InsertPluginIcon } from '../../statics/icons';
+import { InsertPluginIcon } from '../icons';
 
 export default ({ helpers, t }) => {
   return [
@@ -10,7 +10,7 @@ export default ({ helpers, t }) => {
       name: 'Video',
       tooltipText: t('VideoPlugin_InsertButton_Tooltip'),
       Icon: InsertPluginIcon,
-      data: DEFAULTS,
+      componentData: DEFAULTS,
       modalElement: VideoURLInputModal,
       modalStyles: getModalStyles({ fullScreen: false }),
       helpers,